requirements.txt
# JRA中央競馬AI予測モデル Phase 0～5
# Python 3.9 または 3.10 推奨

# ===== 基本ライブラリ =====
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0

# ===== モデリング =====
lightgbm>=4.0.0
optuna>=3.3.0

# ===== 特徴量選択 =====
boruta>=0.3

# ===== データベース接続 =====
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# ===== 可視化 =====
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# ===== ユーティリティ =====
tqdm>=4.65.0
joblib>=1.2.0
pyyaml>=6.0

# ===== 開発用（オプション） =====
jupyter>=1.0.0
ipython>=8.0.0
black>=23.0.0
flake8>=6.0.0

# ===== その他 =====
python-dateutil>=2.8.0
pytz>=2023.3
